requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
Jinja2 